Package command into own package
