+++ src/main/java/laffy/command/AddTodoCommand.java
@@ -1,4 +1,8 @@
-package laffy;
+package laffy.command;
+
+import laffy.Storage;
+import laffy.TaskList;
+import laffy.Ui;
 
 public class AddTodoCommand extends Command {
     public static final String COMMAND_WORD = "todo";
