+++ src/main/java/laffy/command/ListCommand.java
@@ -1,4 +1,8 @@
-package laffy;
+package laffy.command;
+
+import laffy.Storage;
+import laffy.TaskList;
+import laffy.Ui;
 
 public class ListCommand extends Command {
     public static final String COMMAND_WORD = "list";
